extract common base-class for platform loaders, add `$here` alias for use in workers, remove dumb/simpleCache
